+++ inc/packmodules.hpp
@@ -92,5 +92,27 @@ private:
 #endif
 
 
+/**********************
+Number multiplication of matrix.
+Y=kX
+**********************/
+class ProductScalarMatrix: public PackModule
+{
+public:
+    ProductScalarMatrix(Simulator *sim, BusSize size, std::string name="psv");
+    virtual ~ProductScalarMatrix() {}
+    virtual PUnitModule Get_InputPort(int n=0) const override;
+    virtual PMatModule Get_InputBus(int n=0) const override;
+    virtual PMatModule Get_OutputBus(int n=0) const override;
+    zhnmat::Mat Get_OutValue();
+private:
+    BusSize _size;
+    PUnitModule _ugainin=nullptr;
+    PDeMux _dmxin=nullptr;
+    PMux _mxout=nullptr;
+    PUProduct *_prd=nullptr;
+};
+
+
 NAMESPACE_SIMUCPP_R
 #endif  // PACKMODULES_H